requirements.txt
Flask==0.12.2
google-cloud==0.22.0